doc/iso-1939-to-dcat-ap/requirements.txt
beautifulsoup4
lxml
psycopg2
pyyaml
urllib3
